+++ php/tests/Application/Command/SemanticLintTest/UnknownGlobalFunctions.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace A;
+
+function bar()
+{
+
+}
+
+foo();
+bar();
+\A\foo();
+\A\bar();